gpu icp: fixing call to mult kernel

+++ include/libgpualg/mult.cuh
@@ -1,6 +1,6 @@
 #pragma once
 
-void matrixMultiplication(float* A, float* B, float* C,
+void matrixMultiplication(double* A, double* B, double* C,
     int numARows, int numAColumns,
     int numBRows, int numBColumns,
     int numCRows, int numCColumns);
